requirements.txt
asgiref==3.8.1
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.6
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fonttools==4.52.1
fsspec==2024.5.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
